Remove `word_count`
This variable does not means the length of word, it means if select this
word, we should skip how many words behind it.

e.g.
"zhen*" can be translate to "中华儿女", the word length is 4, but if select
it, we should not skip any word behind it.
"woaini" can be translate to "我爱你", if select "我爱", we should skip "ai",
direct show result of "ni"